+++ src/widgets/Navbar/ui/Navbar.module.scss
